Another update 11-03 (2)

+++ NewGame/src/Armor/Armor.java
@@ -65,7 +65,7 @@ public class Armor {
 
         calculateArmorDefence = damageReceived * armorDefencePercentage;
 
-        return calculateArmorDefence;
+        return (int) calculateArmorDefence;
     }
 
     public ArmorCondition getArmorCondition() {
